Move 'register_module_extender' in astroid.brain where it's used
