+++ astroid/__init__.py
@@ -71,23 +71,13 @@ from astroid.scoped_nodes import builtin_lookup
 from astroid.builder import parse, extract_node
 from astroid.util import Uninferable
 
+from astroid.brain.helpers import register_module_extender
+
 # make a manager instance (borg) accessible from astroid package
 from astroid.manager import AstroidManager
 
 MANAGER = AstroidManager()
 del AstroidManager
-
-
-def register_module_extender(manager, module_name, get_extension_mod):
-    def transform(node):
-        extension_module = get_extension_mod()
-        for name, objs in extension_module.locals.items():
-            node.locals[name] = objs
-            for obj in objs:
-                if obj.parent is extension_module:
-                    obj.parent = node
-
-    manager.register_transform(Module, transform, lambda n: n.name == module_name)
 
 
 # load brain plugins
