+++ astroid/brain/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["register_module_extender"]
+
+from astroid.brain.helpers import register_module_extender
