+++ astroid/brain/brain_mechanize.py
@@ -9,7 +9,8 @@
 # Licensed under the LGPL: https://www.gnu.org/licenses/old-licenses/lgpl-2.1.en.html
 # For details: https://github.com/PyCQA/astroid/blob/master/LICENSE
 
-from astroid import MANAGER, register_module_extender
+from astroid import MANAGER
+from astroid.brain.helpers import register_module_extender
 from astroid.builder import AstroidBuilder
