+++ astroid/brain/brain_numpy_core_fromnumeric.py
@@ -9,6 +9,7 @@
 """Astroid hooks for numpy.core.fromnumeric module."""
 
 import astroid
+from astroid.brain.helpers import register_module_extender
 
 
 def numpy_core_fromnumeric_transform():
@@ -20,6 +21,6 @@ def numpy_core_fromnumeric_transform():
     )
 
 
-astroid.register_module_extender(
+register_module_extender(
     astroid.MANAGER, "numpy.core.fromnumeric", numpy_core_fromnumeric_transform
 )
