+++ astroid/brain/brain_numpy_core_umath.py
@@ -12,6 +12,7 @@
 """Astroid hooks for numpy.core.umath module."""
 
 import astroid
+from astroid.brain.helpers import register_module_extender
 
 
 def numpy_core_umath_transform():
@@ -151,6 +152,6 @@ def numpy_core_umath_transform():
     )
 
 
-astroid.register_module_extender(
+register_module_extender(
     astroid.MANAGER, "numpy.core.umath", numpy_core_umath_transform
 )
