+++ astroid/brain/brain_pytest.py
@@ -10,7 +10,8 @@
 # For details: https://github.com/PyCQA/astroid/blob/master/LICENSE
 
 """Astroid hooks for pytest."""
-from astroid import MANAGER, register_module_extender
+from astroid import MANAGER
+from astroid.brain.helpers import register_module_extender
 from astroid.builder import AstroidBuilder
