+++ astroid/brain/brain_ssl.py
@@ -9,7 +9,8 @@
 
 """Astroid hooks for the ssl library."""
 
-from astroid import MANAGER, parse, register_module_extender
+from astroid import MANAGER, parse
+from astroid.brain.helpers import register_module_extender
 
 
 def ssl_transform():
