+++ astroid/brain/brain_threading.py
@@ -6,6 +6,7 @@
 # Licensed under the LGPL: https://www.gnu.org/licenses/old-licenses/lgpl-2.1.en.html
 # For details: https://github.com/PyCQA/astroid/blob/master/LICENSE
 import astroid
+from astroid.brain.helpers import register_module_extender
 
 
 def _thread_transform():
@@ -29,4 +30,4 @@ def _thread_transform():
     )
 
 
-astroid.register_module_extender(astroid.MANAGER, "threading", _thread_transform)
+register_module_extender(astroid.MANAGER, "threading", _thread_transform)
